requirements.txt
# requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.5.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.17.0
openai>=1.0.0
